requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
pymssql==2.3.4